requirements.txt
# Core dependencies
ultralytics==8.0.196
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
easyocr>=1.7.0
Pillow>=9.5.0
numpy>=1.24.0

# Web framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=11.0.2

# Database and encryption
sqlalchemy>=2.0.0
cryptography>=41.0.0
sqlite3-encryption>=0.0.5

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Export formats
onnx>=1.14.0
onnxruntime>=1.16.0
tensorflow>=2.13.0
tf2onnx>=1.15.0

# Utilities
pydantic>=2.4.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
pyyaml>=6.0.1
kaggle>=1.5.16

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Optional mobile dependencies
# coremltools>=7.0  # For iOS Core ML export
# openvino-dev>=2023.1.0  # For Intel optimization